requirements.txt
requests
emoji-country-flag